Database decoding & field reading
