--- Sources/Alchemy/SQL/CodableTools/Decoders/DatabaseDecodingError.swift
@@ -1,4 +1,0 @@
-struct DatabaseDecodingError: Error {
-    let message: String
-    init(_ message: String) { self.message = message }
-}
