+++ Sources/Alchemy/SQL/Database/Abstract/DatabaseError.swift
